🆕 Refactorizacion
